feat(select-git-registry): add create button
