+++ libs/domains/organizations/feature/src/index.ts
@@ -3,6 +3,7 @@ export * from './lib/git-provider-setting/git-provider-setting'
 export * from './lib/git-repository-setting/git-repository-setting'
 export * from './lib/git-token-create-edit-modal/git-token-create-edit-modal'
 export * from './lib/git-token-list/git-token-list'
+export * from './lib/container-registry-create-edit-modal/container-registry-create-edit-modal'
 export * from './lib/hooks/use-add-credit-card/use-add-credit-card'
 export * from './lib/hooks/use-add-credit-code/use-add-credit-code'
 export * from './lib/hooks/use-api-tokens/use-api-tokens'
